requirements.txt
flask
groq
google-generativeai
python-dotenv
gunicorn
requests